everyrhing is working now
