--- src/write/writer.js
@@ -1,7 +1,0 @@
-export function writeFile() {
-
-}
-
-export function writeStd() {
-
-}
